Save Type in ElementTemplate and access it in ShellContent
